backend/integration/src/pyproject.toml
[tool.poetry]
name = "integration"
version = "1.0.0"
description = "Сервис для интеграции данных"
authors = ["vladushka2000"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.32"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
